remove external deps
