+++ Makefile
@@ -243,7 +243,7 @@ all-binaries-platforms:
 # CNI Installer
 .PHONY: azure-cni-manager
 azure-cni-manager: all-binaries-platforms
-	docker build -f ./cni/installer/Dockerfile --build-arg VERSION=$(VERSION) -t $(AZURE_CNI_IMAGE):$(VERSION) .
+	docker build -f ./cni/manager/Dockerfile --build-arg VERSION=$(VERSION) -t $(AZURE_CNI_IMAGE):$(VERSION) .
 
 # Build the Azure CNM plugin image, installable with "docker plugin install".
 .PHONY: azure-vnet-plugin-image
