+++ cni/manager/Dockerfile
@@ -2,7 +2,7 @@ FROM golang as build
 WORKDIR /go/src/github.com/Azure/azure-container-networking/
 ADD . . 
 ARG VERSION
-RUN cd ./cni/installer && CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o /go/installer -ldflags "-X main.version=${VERSION} -s -w"
+RUN cd ./cni/manager && CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o /go/installer -ldflags "-X main.version=${VERSION} -s -w"
 RUN mv ./output /output
 RUN chmod +x /go/installer
 RUN find /output -name "*.zip" -type f -delete
