+++ cni/manager/job-master.yaml
@@ -1,7 +1,7 @@
 apiVersion: batch/v1
 kind: Job
 metadata:
-  name: azure-cni-installer-master
+  name: azure-cni-manager-master
 spec:
   ttlSecondsAfterFinished: 20
   template:
