+++ cni/manager/utils.go
